Fixed #38.
Consolidated all CraftScan options in the chat context menu to a single 'CraftScan' node that expands a sub-menu.

In that new sub-menu, added a context menu option when right clicking a player name in chat to ignore that player in the context of CraftScan. This list is specific to CraftScan matching. You will still see this player's messages in all contexts except CraftScan alerts.