Fixes #25.
Each crafter in the crafter list now has a context menu option for "Local notifications only". When enabled, this option will suppress notifications for that crafter/profession when you are not playing that character.

CraftScan will still generate a match and it will appear in the chat order list, but there will be no notification.