Fixed #49. Keyword matching now prioritizes epic items over other items, assuming all other criteria are the same. If you know the Blue bracer and spark bracer, have a keyword match for 'bracer', then "LF bracer" will match the spark bracer. If a secondary keyword like '629' or 'blue' is configured to lean towards the blue bracer, it will still return the blue bracer.
Fixes #56. Added categoryID persistence. The GetRecipeInfo API returns a 0 categoryID unless you are playing the character with the profession. The first time a profession window is opened after upgrading to the new version, we will walk the list of stored recipes and persist the categoryID so that we have it available to match against category configuration on alts.